Adding in code so that the news updater is run on a cron
